core/standards/fullscreen/interactive/dom-allowfullscreen-attribute-not-set.html

<!DOCTYPE html>
<html>
<head>
	<title>Unset allowfullscreen attribute on iframe</title>
	<meta name="assert" content="To prevent embedded content from going
	fullscreen only embedded content specifically allowed via the allowfullscreen
	attribute of the HTML iframe element will be able to go fullscreen.">
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<style>
		iframe {
			border: none; /* Work-around for Watir bug. */
			width: 400px;
			height: 200px;
		}
	</style>
	<script>
		var test = async_test('Entering fullscreen with allowfullscreen attribute unset');
		var entered_fullscreen;

		function check_if_entered_fullscreen()
		{
			assert_false(entered_fullscreen, "Entered fullscreen");
		}

		function onFullscreenError(e)
		{
			entered_fullscreen = false;
			test.step(check_if_entered_fullscreen);
			test.done();
			removeFullscreenElement();
		}

		document.addEventListener('fullscreenchange', function(e)
		{
			entered_fullscreen = true;
			test.step(check_if_entered_fullscreen)
			test.done();
			removeFullscreenElement();
		}, false);

		function removeFullscreenElement()
		{
			var el = document.fullscreenElement;
			el.parentNode.removeChild(el);
		}
	</script>
</head>
<body>

	<p>Fullscreen mode must not be entered when clicking respective button below.</p>

	<div id="log"></div>

	<iframe src="data:text/html,
		<script>document.addEventListener('fullscreenerror', parent.onFullscreenError, false);</script>
		<button onclick='document.body.requestFullscreen()'>enter fullscreen</button>
		<button onclick='document.exitFullscreen()'>exit fullscreen</button>"></iframe>

</body>
</html>
